agno_service/requirements.txt
phidata==2.6.0
openai==1.40.0
anthropic==0.32.0
google-generativeai==0.7.2
duckduckgo-search==6.2.6
yfinance==0.2.40
pypdf==4.3.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
redis[hiredis]==5.0.7
langfuse==2.42.0
pydantic==2.8.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
